uclogic: Extract report descriptors to a module
